improved historical data processing scripts
